Moved into component
